services/attention-scorer/requirements.txt
grpcio>=1.60.0
numpy>=1.24.0
loguru>=0.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

